A few fixes. Backups finished. - If a task lasted only a few seconds some times it was marked as finished and then everrided by the 100% progress making it imposible to delete. - Added checks for playing, backuping and deleting. - Finished backups system. (Only backup creations, restoration will be implemented soon.) - Improved `compressWorker` progress updater to be more acurate. - Updated es-ES localization.

+++ src/renderer/src/features/installations/pages/ListInstallations.tsx
@@ -19,6 +19,52 @@ function ListInslallations(): JSX.Element {
   const [installationToDelete, setInstallationToDelete] = useState<InstallationType | null>(null)
   const [deleteData, setDeleData] = useState<boolean>(false)
 
+  async function BackupHandler(installation: InstallationType): Promise<void> {
+    if (installation._backuping) return addNotification(t("notifications.titles.error"), t("features.backups.backupInProgress"), "error")
+    if (installation._playing) return addNotification(t("notifications.titles.error"), t("features.backups.backupWhilePlaying"), "error")
+
+    if ((await window.api.pathsManager.checkPathExists(installation.path)) && config.backupsFolder && installation.backupsLimit > 0) {
+      const id = uuidv4()
+      window.api.utils.setPreventAppClose("add", id, "Making and installation backup.")
+
+      configDispatch({ type: CONFIG_ACTIONS.EDIT_INSTALLATION, payload: { id: installation.id, updates: { _backuping: true } } })
+
+      try {
+        let backupsLength = installation.backups.length
+
+        while (backupsLength > 0 && backupsLength >= installation.backupsLimit) {
+          const backupToDelete = installation.backups[backupsLength - 1]
+          const res = await window.api.pathsManager.deletePath(backupToDelete.path)
+          if (!res) return addNotification(t("notifications.titles.error"), t("features.backups.errorDeletingOldBackup"), "error")
+          configDispatch({
+            type: CONFIG_ACTIONS.DELETE_INSTALLATION_BACKUP,
+            payload: { id: installation.id, backupId: backupToDelete.id }
+          })
+          backupsLength--
+          window.api.utils.logMessage("info", `[ListInstallations] [backup] Deleted old backup: ${backupToDelete.path}`)
+        }
+
+        const fileName = `${installation.name.replace(/[^a-zA-Z0-9]/g, "-")}_${new Date().toLocaleString("es").replace(/[^a-zA-Z0-9]/g, "-")}.zip`
+        const backupPath = await window.api.pathsManager.formatPath([config.backupsFolder, "Installations", installation.name.replace(/[^a-zA-Z0-9]/g, "-")])
+        const outBackupPath = await window.api.pathsManager.formatPath([backupPath, fileName])
+
+        await startCompress(`${installation.name} backup`, `Backing up installation ${installation.name}`, installation.path, backupPath, fileName, (status) => {
+          if (!status) throw new Error("Error compressing installation!")
+
+          configDispatch({
+            type: CONFIG_ACTIONS.ADD_INSTALLATION_BACKUP,
+            payload: { id: installation.id, backup: { date: Date.now(), id: uuidv4(), path: outBackupPath } }
+          })
+        })
+      } catch (err) {
+        window.api.utils.logMessage("error", `[ListInstallations] [backup] Error making a backup: ${err}`)
+      } finally {
+        window.api.utils.setPreventAppClose("remove", id, "Finished installation backup.")
+        configDispatch({ type: CONFIG_ACTIONS.EDIT_INSTALLATION, payload: { id: installation.id, updates: { _backuping: false } } })
+      }
+    }
+  }
+
   return (
     <>
       <h1 className="text-3xl text-center font-bold select-none">{t("features.installations.listTitle")}</h1>
@@ -48,40 +94,7 @@ function ListInslallations(): JSX.Element {
                 <Button
                   className="w-7 h-7 bg-zinc-850 shadow shadow-zinc-900 hover:shadow-none flex items-center justify-center rounded"
                   title={t("generic.backup")}
-                  onClick={async () => {
-                    if ((await window.api.pathsManager.checkPathExists(installation.path)) && installation.backupsPath && installation.backupsLimit > 0) {
-                      try {
-                        let backupsLength = installation.backups.length
-
-                        while (backupsLength > 0 && backupsLength >= installation.backupsLimit) {
-                          const backupToDelete = installation.backups[backupsLength - 1]
-                          const res = await window.api.pathsManager.deletePath(backupToDelete.path)
-                          if (!res) return addNotification(t("notifications.titles.error"), "There was an error deleting old backups!", "error")
-                          configDispatch({
-                            type: CONFIG_ACTIONS.DELETE_INSTALLATION_BACKUP,
-                            payload: { id: installation.id, backupId: backupToDelete.id }
-                          })
-                          backupsLength--
-                          window.api.utils.logMessage("info", `[ListInstallations] [backup] Deleted old backup: ${backupToDelete.path}`)
-                        }
-
-                        const fileName = `${installation.name.replace(/[^a-zA-Z0-9]/g, "-")}_${new Date().toLocaleString("es").replace(/[^a-zA-Z0-9]/g, "-")}.zip`
-                        const backupPath = await window.api.pathsManager.formatPath([config.backupsFolder, "Installations"])
-                        const outBackupPath = await window.api.pathsManager.formatPath([backupPath, fileName])
-
-                        await startCompress(`${installation.name} backup`, `Backing up installation ${installation.name}`, installation.path, backupPath, fileName, (status) => {
-                          if (!status) throw new Error("Error compressing installation!")
-
-                          configDispatch({
-                            type: CONFIG_ACTIONS.ADD_INSTALLATION_BACKUP,
-                            payload: { id: installation.id, backup: { date: Date.now(), id: uuidv4(), path: outBackupPath } }
-                          })
-                        })
-                      } catch (err) {
-                        window.api.utils.logMessage("error", `[ListInstallations] [backup] Error making a backup: ${err}`)
-                      }
-                    }
-                  }}
+                  onClick={() => BackupHandler(installation)}
                 >
                   <PiArrowsCounterClockwiseFill className="text-lg" />
                 </Button>
@@ -156,6 +169,9 @@ function ListInslallations(): JSX.Element {
                     className="px-2 py-1 bg-red-800 shadow shadow-zinc-900 hover:shadow-none flex items-center justify-center rounded"
                     onClick={async () => {
                       try {
+                        if (installationToDelete._playing || installationToDelete._backuping)
+                          return addNotification(t("notifications.titles.error"), t("features.installations.cantDeleteWhileinUse"), "error")
+
                         if (deleteData && !(await window.api.pathsManager.deletePath(installationToDelete.path))) throw new Error("Error deleting installation data!")
 
                         configDispatch({ type: CONFIG_ACTIONS.DELETE_INSTALLATION, payload: { id: installationToDelete.id } })
